LayoutTests/http/wpt/webauthn/public-key-credential-get-failure.https-expected.txt

CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.

PASS PublicKeyCredential's [[get]] with timeout
PASS PublicKeyCredential's [[get]] with a mismatched RP ID
PASS PublicKeyCredential's [[get]] with a mismatched APP ID (invalid URLs)
PASS PublicKeyCredential's [[get]] with a mismatched APP ID (different protocols)
PASS PublicKeyCredential's [[get]] with a mismatched APP ID (different sites 1)
PASS PublicKeyCredential's [[get]] with a mismatched APP ID (different sites 2)
PASS PublicKeyCredential's [[get]] with two consecutive requests
PASS PublicKeyCredential's [[get]] with two consecutive requests (2)
PASS PublicKeyCredential's [[get]] with new requests in a new page
PASS PublicKeyCredential's [[get]] with new requests in a new page (2)
